agentic-backend/pyproject.toml
[project]
name = "agentic-backend-app"
readme = "README.md"
version = "1.3.0-post"
description = "Agentic Backend App"
authors = [{ name = "Fred", email = "noreply@example.com" }]
requires-python = ">=3.12,<3.13"

dependencies = [
  "fastapi==0.116.1",
  "fastapi-mcp==0.4.0",
  "PyYAML==6.0.1",
  "uvicorn[standard]>=0.30.6,<0.35.0",
  "croniter==2.0.2",
  "pandas>=2.2.2,<3.0.0",
  "kubernetes==30.1.0",
  "langchain>=1.2.12",
  "langgraph>=1.1.3",
  "langchain-community>=0.4.1",
  "langchain-mcp-adapters>=0.2.1",
  "langfuse>=3.0.0",
  "ipython>=8.30.0,<9.0.0",
  "pyjwt>=2.10.1,<3.0.0",
  "requests>=2.32.3,<3.0.0",
  "pydantic-settings>=2.7.1,<3.0.0",
  "pytest-asyncio>=0.26.0,<0.27.0",
  "python-multipart>=0.0.20,<0.0.23",
  "nest-asyncio>=1.6.0,<2.0.0",
  "opensearch-py>=2.8.0,<3.0.0",
  "rich>=14.0.0",
  "cryptography>=45.0.4",
  "python-pptx>=0.6.23",
  "python-docx>=1.1.0",
  "lxml>=5.3.0",
  "Pillow>=10.0.0",
  "fred-core",
  "uv",
  "pyparsing>=3.2.5",
  "psycopg2-binary>=2.9.10",
  "prometheus-fastapi-instrumentator>=7.1.0",
  "deepagents>=0.4.11",
]

[project.optional-dependencies]
dev = [
  "pylint>=3.3.1,<4.0.0",
  "pip-audit>=2.7.3,<3.0.0",
  "pytest>=8.3.5,<9.0.0",
  "pytest-cov>=6.1.1,<7.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
packages = [{ include = "backend" }]

[tool.setuptools.packages.find]
include = ["fred*"]

[tool.uv.sources]
fred-core = { path = "../fred-core", editable = true }

[dependency-groups]
dev = [
  "bandit>=1.8.6",
  "basedpyright==1.31.0",
  "deepeval>=3.7.0",
  "detect-secrets>=1.5.0",
  "ruff>=0.12.5",
]

[tool.uv]
# Using `editable_mode = "compat"` when installing fred-core to make editable install work with Pylance (vscode):
# https://github.com/microsoft/pylance-release/blob/main/TROUBLESHOOTING.md#editable-install-modules-not-found
config-settings = { editable_mode = "compat" }
